Addressing more issues from merge.

+++ src/expr/proof_rule.cpp
@@ -141,6 +141,6 @@ std::ostream& operator<<(std::ostream& out, PfRule id)
   return out;
 }
 
-size_t PfRuleHashFunction::operator()(PfRule id) const { return id; }
+size_t PfRuleHashFunction::operator()(PfRule id) const { return static_cast<size_t>(id); }
 
 }  // namespace CVC4
